feat: creating request validation
