+++ src/routes.py
@@ -1,6 +1,7 @@
 from src.Application.Controllers.user_controller import UserController
 from src.Application.Controllers.seller_controler import SellerController
-from flask import jsonify, make_response
+from flask import jsonify, make_response, request
+from src.Application.Dto.seller_dto import SellerRegisterSchema
 
 def init_routes(app):    
     @app.route('/api', methods=['GET'])
@@ -14,7 +15,8 @@ def init_routes(app):
         return UserController.register_user()
     
     @app.route('/seller', methods=['POST'])
-    def register_seller():
-        return SellerController.register_seller()
+    @app.input(SellerRegisterSchema, location='json')
+    def register_seller(body):
+        return SellerController.register_seller(body)
